Update WCS to support FK4 B1950 and all-sky projections
- Add FK4 B1950 frame transforms with E-terms of aberration, validated
  against astropy to ~mas accuracy; FK4 at other equinoxes rejected
  cleanly
- Implement AIT (Hammer–Aitoff) and MOL (Mollweide) all-sky projections
- Include IAU-2000 frame bias in FK5 transforms (now ~25 mas accurate vs
  ICRS)
- Update CLAUDE.md and ROADMAP.md to reflect completion of Phase 1
  binary-table
  VLA write and Phase 6 WCS v2 feature set
- Minor formatting fixes in compress tests

+++ src/compress/tests.rs
@@ -353,7 +353,7 @@ fn hcompress_lossy_write_round_trips_within_scale() {
 
     // A smooth-ish 32×32 i32 image; lossy HCOMPRESS with SCALE=4.
     let samples: Vec<i32> = (0..32 * 32)
-        .map(|i| (100 + 5 * (i % 32) + 3 * (i / 32)) as i32)
+        .map(|i| 100 + 5 * (i % 32) + 3 * (i / 32))
         .collect();
     let image = Image {
         shape: vec![32, 32],
@@ -379,7 +379,10 @@ fn hcompress_lossy_write_round_trips_within_scale() {
         .map(|(a, b)| (a - b).abs())
         .max()
         .unwrap();
-    assert!(max_err <= 4, "HCOMPRESS lossy error {max_err} exceeds scale");
+    assert!(
+        max_err <= 4,
+        "HCOMPRESS lossy error {max_err} exceeds scale"
+    );
 }
 
 #[test]
